Moved controller into game folder
