+++ src/components/game/control/GameController.js
@@ -1,4 +1,4 @@
-import { Angle, Dxn } from "../game/coreState/utils/Direction";
+import { Angle, Dxn } from "../coreState/utils/Direction";
 import { GameAction, ActionType } from "./GameAction";
 // A class whose instance acts as a UseState for canvas to listen and hold onto keystrokes, to be consumed by a GameState on its update.
